Various issues
Fixed remaining trailing stop.  When the bot's trailing stop triggered
it dumps all of that coin and waits to buy.  The problem is that when it
sold everything, the trailing stop remained in the database and would
trigger sells for 0 amount.  Now, when the current holding quantity is 0
the trailing stop is deleted and ready for the next purchase.

Added a last stop filter to the buy and sell functions to prevent
transactions with valuations of less than .01 BTC.

Reduced the amount of allowed open orders.  For all trading strategies
only one open order is allowed.  The bot will wait for that order to be
filled, or until the timeout (2 periods) is reached and it cancels the
order itself.  I will add some logic so it will only post one order, but
will gauge all other orders by the posted order.  Will always post the
best order available to catch the most recent action.  This feature will
come soon.

Fixed issue with double emails being sent for each transaction.   My
personal favorite fix this cycle.